backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
supabase==2.3.4
sqlalchemy==2.0.23
pydantic==2.4.2
python-multipart==0.0.6
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.26.2
mediapipe==0.10.5
boto3==1.29.7
aiofiles==23.2.1
pytest==7.4.3
httpx==0.25.0
